requirements.txt
requests
python-dotenv
pytz
boto3
matplotlib

